+++ crates/ryo3/src/dev/sp/pydone.rs
@@ -1,7 +1,7 @@
 use pyo3::prelude::*;
 use pyo3::types::PyBytes;
 
-use crate::sp::done::Done;
+use crate::dev::sp::done::Done;
 
 #[pyclass]
 #[derive(Debug)]
@@ -28,6 +28,21 @@ impl PyDone {
     #[getter]
     fn stdout<'py>(&'py self, py: Python<'py>) -> Bound<'py, PyBytes> {
         PyBytes::new_bound(py, &self.done.stdout)
+    }
+
+    #[getter]
+    fn stderr<'py>(&'py self, py: Python<'py>) -> Bound<'py, PyBytes> {
+        PyBytes::new_bound(py, &self.done.stderr)
+    }
+
+    #[getter]
+    fn returncode(&self) -> i32 {
+        self.done.returncode
+    }
+
+    #[getter]
+    fn args(&self) -> Vec<String> {
+        self.done.args.clone()
     }
 }