+++ crates/ryo3/src/fs/fileio.rs
@@ -1,18 +1,18 @@
 use std::path::Path;
 
-use pyo3::exceptions::{PyFileNotFoundError, PyUnicodeDecodeError};
+use pyo3::exceptions::{PyFileNotFoundError, PyNotADirectoryError, PyUnicodeDecodeError};
 use pyo3::prelude::*;
 use pyo3::types::{PyBytes, PyModule};
 use pyo3::{pyfunction, wrap_pyfunction, PyResult};
 
 #[pyfunction]
 pub fn read_vec_u8(s: &str) -> PyResult<Vec<u8>> {
-    let p = Path::new(s);
-    let b = std::fs::read(p);
-    match b {
+    let fpath = Path::new(s);
+    let fbytes = std::fs::read(fpath);
+    match fbytes {
         Ok(b) => Ok(b),
         Err(e) => {
-            let emsg = format!("{}: {} - {:?}", p.to_str().unwrap(), e, p.to_str().unwrap());
+            let emsg = format!("read_vec_u8 - path: {s} - {e}");
             Err(PyFileNotFoundError::new_err(emsg))
         }
     }
@@ -31,35 +31,33 @@ pub fn read_text(py: Python<'_>, s: &str) -> PyResult<String> {
     match r {
         Ok(s) => Ok(s.to_string()),
         Err(e) => {
-            let decode_err = PyUnicodeDecodeError::new_utf8_bound(py, &bvec, e).unwrap();
+            let decode_err = PyUnicodeDecodeError::new_utf8_bound(py, &bvec, e)?;
             Err(decode_err.into())
         }
     }
 }
 
 #[pyfunction]
-pub fn write_bytes(s: &str, b: Vec<u8>) -> PyResult<()> {
-    let p = Path::new(s);
-    let r = std::fs::write(p, b);
-    match r {
-        Ok(_) => Ok(()),
-        Err(e) => {
-            let emsg = format!("{}: {} - {:?}", p.to_str().unwrap(), e, p.to_str().unwrap());
-            Err(PyFileNotFoundError::new_err(emsg))
-        }
+pub fn write_bytes(fspath: &str, b: Vec<u8>) -> PyResult<()> {
+    let fpath = Path::new(fspath);
+    let write_res = std::fs::write(fpath, b);
+    match write_res {
+        Ok(()) => Ok(()),
+        Err(e) => Err(PyNotADirectoryError::new_err(format!(
+            "write_bytes - parent: {fspath} - {e}"
+        ))),
     }
 }
 
 #[pyfunction]
-pub fn write_text(s: &str, t: &str) -> PyResult<()> {
-    let p = Path::new(s);
-    let r = std::fs::write(p, t);
-    match r {
-        Ok(_) => Ok(()),
-        Err(e) => {
-            let emsg = format!("{}: {} - {:?}", p.to_str().unwrap(), e, p.to_str().unwrap());
-            Err(PyFileNotFoundError::new_err(emsg))
-        }
+pub fn write_text(fspath: &str, string: &str) -> PyResult<()> {
+    let fpath = Path::new(fspath);
+    let write_result = std::fs::write(fpath, string);
+    match write_result {
+        Ok(()) => Ok(()),
+        Err(e) => Err(PyNotADirectoryError::new_err(format!(
+            "write_bytes - parent: {fspath} - {e}"
+        ))),
     }
 }