+++ crates/ryo3/src/libs/which.rs
@@ -6,21 +6,26 @@ use pyo3::prelude::*;
 
 #[pyfunction]
 pub fn which(cmd: &str, path: Option<&str>) -> PyResult<Option<std::path::PathBuf>> {
-    match path {
-        Some(p) => {
-            let which_res = which_rs::which_in(cmd, Some(p), env::current_dir().unwrap());
-            match which_res {
-                Ok(p) => Ok(Some(p)),
-                Err(_e) => Ok(None),
+    if let Some(p) = path {
+        // get current directory w/o unwrapping
+        match env::current_dir() {
+            Ok(c) => {
+                let which_res = which_rs::which_in(cmd, Some(p), c);
+                match which_res {
+                    Ok(p) => Ok(Some(p)),
+                    Err(_e) => Ok(None),
+                }
             }
+            Err(_e) => Err(PyErr::new::<pyo3::exceptions::PyOSError, _>(
+                "which: current directory is not a valid path",
+            )),
         }
-        None => {
-            let r = which_rs::which(cmd);
+    } else {
+        let r = which_rs::which(cmd);
 
-            match r {
-                Ok(p) => Ok(Some(p)),
-                Err(_e) => Ok(None),
-            }
+        match r {
+            Ok(p) => Ok(Some(p)),
+            Err(_e) => Ok(None),
         }
     }
 }
@@ -31,10 +36,23 @@ pub fn which_all(cmd: &str, path: Option<&str>) -> PyResult<Vec<String>> {
         Some(p) => Some(OsString::from(p)),
         None => env::var_os("PATH"),
     };
-    let which_iter = which_rs::which_in_all(cmd, search_path, env::current_dir().unwrap()).unwrap();
+    let which_iter = which_rs::which_in_all(
+        cmd,
+        search_path,
+        env::current_dir().expect("which_all: current directory is not a valid path"),
+    )
+    .map_err(|_| {
+        PyErr::new::<pyo3::exceptions::PyOSError, _>(
+            "which_all: current directory is not a valid path",
+        )
+    })?;
     let which_vec = which_iter
         .into_iter()
-        .map(|p| p.to_str().unwrap().to_string())
+        .map(|p| {
+            p.to_str()
+                .expect("which_all: path contains invalid unicode characters")
+                .to_string()
+        })
         .collect::<Vec<String>>();
     Ok(which_vec)
 }