+++ tests/test_run.py
@@ -10,14 +10,13 @@ def test_python_version() -> None:
     assert completed_proc.returncode == 0
     print(completed_proc)
     assert "python" in str(completed_proc.stdout).lower()
-
-    thing = ry.run(
+    res = ry.run(  # type: ignore
         python_exe,
         "--version",
-        # capture_output=True,
     )
-    print(thing)
-    # assert False
+    print(res)
+    print(dir(res))
+    assert res.returncode == 0
 
 
 def test_binary_output() -> None: